+++ src/main/kotlin/com/jetbrains/rider/aspire/actions/StartAspireWorkerAction.kt
@@ -5,7 +5,7 @@ package com.jetbrains.rider.aspire.actions
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.progress.currentThreadCoroutineScope
-import com.jetbrains.rider.aspire.sessionHost.AspireWorkerManager
+import com.jetbrains.rider.aspire.worker.AspireWorkerManager
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
