+++ src/main/kotlin/com/jetbrains/rider/aspire/worker/AspireWorkerManager.kt
@@ -1,6 +1,4 @@
-@file:Suppress("UnstableApiUsage")
-
-package com.jetbrains.rider.aspire.sessionHost
+package com.jetbrains.rider.aspire.worker
 
 import com.intellij.execution.RunManager
 import com.intellij.execution.RunManagerListener
@@ -61,7 +59,7 @@ class AspireWorkerManager(private val project: Project, scope: CoroutineScope) :
 
     fun removeAspireHost(projectFilePath: String) {
         val configurationType = ConfigurationTypeUtil.findConfigurationType(AspireHostConfigurationType::class.java)
-        val configurations = RunManager.getInstance(project)
+        val configurations = RunManager.Companion.getInstance(project)
             .getConfigurationsList(configurationType)
             .filter { it is AspireHostConfiguration && it.parameters.projectFilePath == projectFilePath }
         if (configurations.isNotEmpty()) return
